feat(change-email): Updated to move the location of the command and renamed.
